api/pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "Basic Todo application"
authors = ["Mason Preston <z3r0.mp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
pymongo = "^4.6.2"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core>=1.6"]
build-backend = "poetry.core.masonry.api"
